feat: allow users to create the same keyword but with different parameters

+++ bot/src/commands/keyword/keyword-delete.sub.ts
@@ -4,6 +4,7 @@ import PrismaErrorCode from '../../util/prisma-error-code';
 import KeywordCache from '../../service/keyword-cache';
 import BotError from '../../util/bot-error';
 import prisma from '../../util/prisma';
+import { KeywordAction } from '@prisma/client';
 
 const KeywordDelete: Command = {
     id: 'f10df570-b711-41bc-a083-0b2465304b1e',
@@ -11,8 +12,32 @@ const KeywordDelete: Command = {
     handler: async (interaction, user, server) => {
         const word = interaction.options.getString('keyword', true);
         const scoreName = interaction.options.getString('score-name', true);
+        const action = interaction.options.getString('action');
+        const amount = interaction.options.getInteger('amount');
+        const channel = interaction.options.getChannel('channel');
+        const filterOnUser = interaction.options.getUser('user');
+        const role = interaction.options.getRole('role');
+
+
+        let inDbUser;
+        try {
+            if (filterOnUser) {
+                inDbUser = await prisma.user.findUnique({
+                    where: {
+                        discordId: user.id
+                    }
+                });
+            }
+        } catch (e) {
+            if ((e as PrismaClientKnownRequestError).code === PrismaErrorCode.NOT_FOUND)
+                throw new BotError('Could not find user to filter on.');
+        }
 
         try {
+            if (!server)
+                throw new Error(`Server instance not instantiated. Keyword functionality has been disabled.`);
+
+
             const score = await prisma.score.findUnique({
                 where: {
                     name_serverId: {
@@ -22,20 +47,31 @@ const KeywordDelete: Command = {
                 }
             });
     
-            const keywordRecord = await prisma.keyword.delete({
+            const keywordRecord = await prisma.keyword.findFirst({
                 where: {
-                    word_scoreId_serverId: {
-                        word,
-                        scoreId: score.id,
-                        serverId: server?.id as string,
-                    }
+                    word,
+                    scoreId: score?.id as string,
+                    serverId: server?.id as string,
+                    channelId: channel?.id,
+                    amount: amount || undefined,
+                    action: action as KeywordAction || undefined,
+                    userId: inDbUser?.id,
+                    roleId: role?.id
+                },
+                rejectOnNotFound: false
+            }); 
+
+            if (!keywordRecord)
+                throw new BotError(`The keyword does not exist.`); 
+
+            await prisma.keyword.delete({
+                where: {
+                    id: keywordRecord.id
                 }
             });
     
             await KeywordCache.deleteRecord(server.id, keywordRecord.id);
-        } catch (e) {
-            if ((e as PrismaClientKnownRequestError).code === PrismaErrorCode.NOT_FOUND)
-                throw new BotError(`The keyword/score does not exist.`);
+        } catch (e) {                
             throw e;
         }
